feat: modal container create, chat modal refactoring
